Gestion de memoria con placement new y uso de plantillas para acceder a los componentes. Falta usar una plantilla para crearlos

+++ src/cmp/AI.hpp
@@ -8,7 +8,7 @@
 
 using Sun::Vector3f;
 
-constexpr static unsigned short MAX_PATROL_POSITIONS = 6;
+constexpr static unsigned int MAX_PATROL_POSITIONS = 6;
 
 struct AI : public Component {
 	explicit AI(EntityType e_type, std::size_t e_ID, const std::vector<Vector3f>& patrol);
@@ -17,10 +17,10 @@ struct AI : public Component {
 
 	AI_State state { PATROL_STATE };
 
-	Vector3f patrol_position[MAX_PATROL_POSITIONS];
+	Vector3f patrol_position[MAX_PATROL_POSITIONS]; // intentar hacer const
 
 	unsigned short patrol_index { 0 };
-	unsigned short max_index { 0 };
+	unsigned short max_index { 0 };		// intentar hacer const
 
 	Vector3f target_position {0};
 };
