+++ src/cmp/Component.cpp
@@ -2,8 +2,8 @@
 
 std::ostream& Component::print(std::ostream& os, const std::string& name) const {
 	return 	os	<< "\n\n"		<< name
-				<< " ( e_ID = " << entityID
-				<< " c_ID = " 	<< ID
-				<< " e_Type = " << entityType
+				<< " ( e_ID = " << static_cast<uint32_t>(entityID)
+				<< " c_ID = " 	<< static_cast<uint32_t>(ID)
+				<< " e_Type = " << static_cast<uint32_t>(entityType)
 				<< " )";
 }
